Problem: changes are in unexpected directory
Solution: move to scripts/DMF
